更新 WorkWithHybridCLR项目，与最新 obfuz4hybridclr的代码同步

+++ Samples/WorkWithHybridCLR/Assets/Editor/BuildCommand.cs
@@ -1,4 +1,6 @@
 using HybridCLR.Editor;
+using HybridCLR.Editor.Commands;
+using Obfuz.Settings;
 using Obfuz4HybridCLR;
 using System.Collections;
 using System.Collections.Generic;
@@ -12,17 +14,26 @@ public static class BuildCommand
     public static void CompileAndObfuscateAndCopyToStreamingAssets()
     {
         BuildTarget target = EditorUserBuildSettings.activeBuildTarget;
-        ObfuscateUtil.CompileAndObfuscateHotUpdateAssemblies(target);
+        CompileDllCommand.CompileDll(target);
+
+        string obfuscatedHotUpdateDllPath = PrebuildCommandExt.GetObfuscatedHotUpdateAssemblyOutputPath(target);
+        ObfuscateUtil.ObfuscateHotUpdateAssemblies(target, obfuscatedHotUpdateDllPath);
 
         Directory.CreateDirectory(Application.streamingAssetsPath);
 
         string hotUpdateDllPath = $"{SettingsUtil.GetHotUpdateDllsOutputDirByTarget(target)}";
+        List<string> obfuscationRelativeAssemblyNames = ObfuzSettings.Instance.assemblySettings.GetObfuscationRelativeAssemblyNames();
+
         foreach (string assName in SettingsUtil.HotUpdateAssemblyNamesIncludePreserved)
         {
-            string srcFile = $"{hotUpdateDllPath}/{assName}.dll";
+            string srcDir = obfuscationRelativeAssemblyNames.Contains(assName) ? obfuscatedHotUpdateDllPath : hotUpdateDllPath;
+            string srcFile = $"{srcDir}/{assName}.dll";
             string dstFile = $"{Application.streamingAssetsPath}/{assName}.dll.bytes";
-            File.Copy(srcFile, dstFile, true);
-            Debug.Log($"[CompileAndObfuscate] Copy {srcFile} to {dstFile}");
+            if (File.Exists(srcFile))
+            {
+                File.Copy(srcFile, dstFile, true);
+                Debug.Log($"[CompileAndObfuscate] Copy {srcFile} to {dstFile}");
+            }
         }
     }
 }
